+++ dorado/__init__.py
@@ -4,26 +4,25 @@
 # should keep this content at the top.
 # ----------------------------------------------------------------------------
 from ._astropy_init import *   # noqa
+
+from .zellars import *
+from .ceres import *
+from .stack import *
+from .clippy import *
+from .fournax import *
+from .timeseries import *
 # ----------------------------------------------------------------------------
 
-__all__ = []
+# __all__ = []
+# from .config import *
 # from .example_mod import *   # noqa
-from .fournax import *
-from .vizual import *
-from .rchive import *
-from .waldo import *
-from .imreduc import *
-from .lazyeye import *
 
-__all__ += fournax.__all__
-__all__ += vizual.__all__
-__all__ += rchive.__all__
-__all__ += waldo.__all__
-__all__ += imreduc.__all__
-__all__ += lazyeye.__all__
-# __all__ += dorado.__all__
 
 # Then you can be explicit to control what ends up in the namespace,
 # __all__ += ['do_primes']   # noqa
 # or you can keep everything from the subpackage with the following instead
 # __all__ += example_mod.__all__
+
+
+# .. warning::
+# :mod:`word`