+++ dorado/ceres/__init__.py
@@ -1,0 +1,9 @@
+"""
+This is the docstring for the examplesubpkg package.  Normally you would
+have whatever.py files in this directory implementing some modules, but this
+is just an example sub-package, so it doesn't actually do anything.
+"""
+__all__ = []
+
+from .ceresClass import *
+__all__ += ceresClass.__all__