+++ dorado/config/config.py
@@ -1,0 +1,4 @@
+from __init__ import conf
+
+print(conf.use_logger)
+print(conf.get_config)