+++ dorado/fournax/__init__.py
@@ -1,9 +1,5 @@
-"""
-This is the docstring for the examplesubpkg package.  Normally you would
-have whatever.py files in this directory implementing some modules, but this
-is just an example sub-package, so it doesn't actually do anything.
-"""
 __all__ = []
 
-from .fournax import *
-__all__ += fournax.__all__
+from .fournaxClass import *
+
+__all__ += fournaxClass.__all__