+++ dorado/stack/__init__.py
@@ -1,0 +1,6 @@
+
+__all__ = []
+
+from .stackClass import *
+
+__all__ += stackClass.__all__