+++ dorado/timeseries/__init__.py
@@ -1,0 +1,5 @@
+__all__ = []
+
+from .timeseriesClass import *
+
+__all__ += timeseriesClass.__all__