+++ dorado/zellars/__init__.py
@@ -1,0 +1,4 @@
+__all__ = []
+
+from .zellarsClass import *
+__all__ += zellarsClass.__all__